gui/requirements.txt
PySide6>=6.5.0
psutil>=5.9.0
PyNaCl>=1.5.0
curl-cffi>=0.7.0
pyobjc-framework-Cocoa>=9.0; sys_platform == "darwin"
